Pathfinding Commit #1
This was meant to have been uploaded much earlier in the week!
Very sorry!
[+] Added the ability for enemies to navigate any level via cleverly
placed "waypoint tiles" that contain directional information for enemies
to make use of
[+] Re-arranged what numbers mean what in the array that is used to load
tiles into the game, I did this because originally the numbers started
as single digits then went into double digits, however the map becomes
very difficult to make sense of when the numbers are pushing lines out
of sync with on another